pyproject.toml
[project]
name = "journal-prompts"
version = "0.1.0"
description = "Daily journaling prompt generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0",
  "openai>=1.40.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
